Jwt Authentication of the Api

+++ backend/controllers/getController.js
@@ -2,11 +2,13 @@ const asyncHandler = require('express-async-handler');
 
 //bring in the Goal Model
 const Goal = require('../models/goalModel');
+//User Model
+const User = require('../models/userModel');
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -20,6 +22,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -32,6 +35,22 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('No Goal Found');
   }
+
+  //..get user
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User Not found');
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User Not authorized');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -46,6 +65,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('No Goal Found');
+  }
+  //..get user
+  const user = await User.findById(req.user.id);
+
+  //check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User Not found');
+  }
+
+  //make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User Not authorized');
   }
   //remove the goal
   await goal.deleteOne();
